refactor(SkeletonItem): 반응형에도 잘 동작하도록 로직 수정
