+++ src/components/SkeletonItem/skeleton-item.css
@@ -23,10 +23,6 @@
   border-radius: 8px;
 }
 
-.skeleton-list {
-  margin-bottom: 48px;
-}
-
 .hide-skeleton {
   display: none !important;
 }
